// apps/api/src/common/errors/business-error-code.ts
export enum BusinessErrorCode {
  EMAIL_JA_EM_USO = 'EMAIL_JA_EM_USO',
  TOKEN_VERIFICACAO_INVALIDO = 'TOKEN_VERIFICACAO_INVALIDO',
  TOKEN_VERIFICACAO_EXPIRADO = 'TOKEN_VERIFICACAO_EXPIRADO',
  TOKEN_REDEFINICAO_INVALIDO = 'TOKEN_REDEFINICAO_INVALIDO',
  TOKEN_REDEFINICAO_EXPIRADO = 'TOKEN_REDEFINICAO_EXPIRADO',
  CREDENCIAIS_INVALIDAS = 'CREDENCIAIS_INVALIDAS',
  EMAIL_NAO_VERIFICADO = 'EMAIL_NAO_VERIFICADO',
  SESSAO_INVALIDA = 'SESSAO_INVALIDA',
  MUITAS_TENTATIVAS = 'MUITAS_TENTATIVAS',
  LAR_NAO_ENCONTRADO = 'LAR_NAO_ENCONTRADO',
  USUARIO_NAO_ENCONTRADO = 'USUARIO_NAO_ENCONTRADO',
  APENAS_ANFITRIAO = 'APENAS_ANFITRIAO',
  PLANO_LIMITE_LARES = 'PLANO_LIMITE_LARES',
  PLANO_LIMITE_MEMBROS = 'PLANO_LIMITE_MEMBROS',
  PLANO_LIMITE_PARTICIPACAO = 'PLANO_LIMITE_PARTICIPACAO',
  CONVITE_INVALIDO = 'CONVITE_INVALIDO',
  CONVITE_EXPIRADO = 'CONVITE_EXPIRADO',
  CONVITE_JA_UTILIZADO = 'CONVITE_JA_UTILIZADO',
  CONVITE_DUPLICADO = 'CONVITE_DUPLICADO',
  USUARIO_JA_NO_LAR = 'USUARIO_JA_NO_LAR',
  SEM_PERMISSAO_DOCUMENTO = 'SEM_PERMISSAO_DOCUMENTO',
  PLANO_LIMITE_DOCUMENTOS = 'PLANO_LIMITE_DOCUMENTOS',
  MEMBRO_DOCUMENTO_INVALIDO = 'MEMBRO_DOCUMENTO_INVALIDO',
  SEM_PERMISSAO_EVENTO = 'SEM_PERMISSAO_EVENTO',
  MEMBRO_EVENTO_INVALIDO = 'MEMBRO_EVENTO_INVALIDO',
  EVENTO_PERIODO_INVALIDO = 'EVENTO_PERIODO_INVALIDO',
  ARQUIVO_INVALIDO = 'ARQUIVO_INVALIDO',
  DOCUMENTO_ARQUIVO_INDISPONIVEL = 'DOCUMENTO_ARQUIVO_INDISPONIVEL',
  RATE_LIMIT_ARQUIVOS = 'RATE_LIMIT_ARQUIVOS',
  PLANO_LIMITE_AGENDA = 'PLANO_LIMITE_AGENDA',
}